figured out how to refactor styles effeciently
